organize pt2
